# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: La Escoba Card Game API
  description: Microservices backend for the Spanish card game "La Escoba".
  version: 1.0.0
  contact:
    name: ASE Project Team
servers:
  - url: https://localhost:5000
    description: Local Development Server (HTTPS Gateway)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /matches:
    post:
      summary: Create a new match
      tags: [Match]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player1: { type: string }
                player2: { type: string }
      responses:
        '201':
          description: Match created
          content:
            application/json:
              schema:
                type: object
                properties:
                  match_id: { type: string }
                  turn: { type: string }
  /matches/{match_id}:
    get:
      summary: Get match state (hand and table)
      tags: [Match]
      parameters:
        - in: path
          name: match_id
          required: true
          schema: { type: string }
        - in: query
          name: player
          required: true
          schema: { type: string }
          description: The player requesting the state (to hide opponent cards)
      responses:
        '200':
          description: Current match state
          content:
            application/json:
              schema:
                type: object
                properties:
                  table: { type: array, items: { type: integer } }
                  status: { type: string }
                  turn: { type: string }
  /matches/{match_id}/play:
    post:
      summary: Play a card (move)
      tags: [Match]
      parameters:
        - in: path
          name: match_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player: { type: string }
                card_id: { type: integer }
      responses:
        '200':
          description: Move accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  captured: { type: array, items: { type: integer } }
                  escoba: { type: boolean }

  # --- HISTORY SERVICE ---
...